Updated the order of the parameters for logistic and reg_logistic
As required in the project description, the function [reg] logistic should have a certain order of parametrs, I changed batch_size to the last parameter so it does not interfer with the automatic grading system.

I also deleted "learning_by_stoch_gradient_descent" which was logistic with SGD. It was not used anymore since a while, because it has been combined with logistic_regression. Simply set batch_size appropriately.